refactor: Move TestGitServer into git package
