+++ pkg/git/test_git_server.go
@@ -1,4 +1,4 @@
-package test_utils
+package git
 
 import (
 	"context"
