+++ pkg/vars/vars_loader_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	git2 "github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing"
-	"github.com/kluctl/kluctl/v2/e2e/test-utils"
+	"github.com/kluctl/kluctl/v2/pkg/git"
 	"github.com/kluctl/kluctl/v2/pkg/git/auth"
 	git_url "github.com/kluctl/kluctl/v2/pkg/git/git-url"
 	ssh_pool "github.com/kluctl/kluctl/v2/pkg/git/ssh-pool"
@@ -155,7 +155,7 @@ func TestVarsLoader_FileWithLoadNotExists(t *testing.T) {
 }
 
 func TestVarsLoader_Git(t *testing.T) {
-	gs := test_utils.NewTestGitServer(t)
+	gs := git.NewTestGitServer(t)
 	gs.GitInit("repo")
 	gs.UpdateYaml("repo", "test.yaml", func(o *uo.UnstructuredObject) error {
 		*o = *uo.FromStringMust(`{"test1": {"test2": 42}}`)
@@ -178,7 +178,7 @@ func TestVarsLoader_Git(t *testing.T) {
 }
 
 func TestVarsLoader_GitBranch(t *testing.T) {
-	gs := test_utils.NewTestGitServer(t)
+	gs := git.NewTestGitServer(t)
 	gs.GitInit("repo")
 
 	wt := gs.GetWorktree("repo")
